requirements.txt
findspark==2.0.1
joblib==1.3.2
kafka_python==2.0.2
numpy==1.25.2
pandas==2.2.2
plotly==5.18.0
Requests==2.32.3
selenium==4.23.1
streamlit==1.35.0
webdriver_manager==4.0.1
